MG: Make it possible to disable SD card support (I don't use this)
Mostly a lot of "#if ENABLED(SDSUPPORT) wrapping.

For this pass I've left most of the SD function calls in but gutted them
so they do nothing or log an error. The compiler SHOULD optimize them away.

TODO: Actually remove the functions so they're not polluting the namespace.
(Also frees program/data memory if the compiler isn't optimizing right)